backend/requirements.txt
fastapi
uvicorn
aiosqlite
openai
python-dotenv
pydantic
python-multipart